community_dev/web_app/requirements.txt
python-fasthtml>=0.1.0
python-multipart>=0.0.6
uvicorn>=0.15.0
pandas>=2.1.0
numpy>=1.24.0
loguru>=0.7.2
pydantic>=2.0.0
click>=8.1.0
rich>=13.0.0 